MVVM
Trying to do MVVM but it seems that I’m breaking a bunch while doing
so. I made this update prematurely so that in case anyone wants to
tweak something they can do it on the proper file
