+++ Health/src/main/res/layout/content_main.xml
@@ -9,8 +9,8 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.example.sciencesquad.health.HealthMainActivity"
-    tools:showIn="@layout/app_bar_health_main">
+    tools:context=".MainActivity"
+    tools:showIn="@layout/app_bar_main">
 
     <TextView
         android:layout_width="wrap_content"
